+++ src/main/java/com/qa/calculator/Editable.java
@@ -2,5 +2,5 @@ package com.qa.calculator;
 
 public interface Editable {
 
-	public int editNumbers(int num1, int num2);
+	public double editNumbers(int num1, int num2);
 }
